src/main/java/hudson/plugins/sctmexecutor/Utils.java
package hudson.plugins.sctmexecutor;

import java.util.LinkedList;
import java.util.List;

public final class Utils {
  public static List<Integer> csvToIntList(String execDefIds) {
    List<Integer> list = new LinkedList<Integer>();
    if (execDefIds.contains(",")) { //$NON-NLS-1$
      String[] ids = execDefIds.split(","); //$NON-NLS-1$
      for (String str : ids) {
        list.add(Integer.valueOf(str));
      }
    } else {
      list.add(Integer.valueOf(execDefIds));
    }
    return list;
  }
}
